requirements.txt
fastapi
uvicorn
psycopg2-binary
langchain
langchain-groq
python-dotenv
python-multipart
